corregi las rutas

+++ src/controllers/pedidos.controllers.js
@@ -43,6 +43,12 @@ export const consultaEditarPedido = async(req, res)=>{
 
 export const consultaBorrarPedido = async(req, res)=>{
     try {
+        const pedidoABorrar = await Pedido.findById(req.params.id);
+        if(!pedidoABorrar){
+            return res.status(404).json({
+                mensaje: "El pedido que querés borrar no fue encontrado.",
+            });
+        }
         await Pedido.findByIdAndDelete(req.params.id);
         res.status(200).json({
             mensaje: 'El pedido fue eliminado correctamente',
